requirements.txt
# packaging
pip
setuptools
twine
wheel
colorama
packaging<24
pyarrow
yapf
msgpack

# data science
numpy
pandas
scikit-learn

# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# Images
opencv-python

# tests/linter
pytest
pylint

# Google CLoud Platform
google-cloud-storage
google-cloud-bigquery
db-dtypes

# MLOps
mlflow
prefect
psycopg2-binary

# API
fastapi
pytz
uvicorn
protobuf==3.20.*

# tests
httpx           # HTTP client
pytest-asyncio  # asynchronous I/O support for pytest

# Utilities
